test(use): loop over tservers
